ML/requirements.txt
tensorflow
opencv-python
pytesseract
flask
flask_cors
pillow
gunicorn